+++ packages/auth-client/src/index.ts
@@ -28,7 +28,13 @@ import {
 } from './storage';
 import { PartialIdentity } from '@dfinity/identity/lib/cjs/identity/partial';
 
-export { AuthClientStorage, IdbStorage, LocalStorage, KEY_STORAGE_DELEGATION, KEY_STORAGE_KEY } from './storage';
+export {
+  AuthClientStorage,
+  IdbStorage,
+  LocalStorage,
+  KEY_STORAGE_DELEGATION,
+  KEY_STORAGE_KEY,
+} from './storage';
 export { IdbKeyVal, DBCreateOptions } from './db';
 
 const IDENTITY_PROVIDER_DEFAULT = 'https://identity.ic0.app';
@@ -429,7 +435,11 @@ export class AuthClient {
   }
 
   public async isAuthenticated(): Promise<boolean> {
-    return !this.getIdentity().getPrincipal().isAnonymous() && this._chain !== null;
+    return (
+      !this.getIdentity().getPrincipal().isAnonymous() &&
+      this._chain !== null &&
+      isDelegationValid(this._chain)
+    );
   }
 
   /**